Move tap flag into separate viz profile command
