+++ pkg/profiles/proto_test.go
@@ -33,7 +33,7 @@ service VotingService {
 	parser := proto.NewParser(strings.NewReader(protobuf))
 
 	expectedServiceProfile := sp.ServiceProfile{
-		TypeMeta: serviceProfileMeta,
+		TypeMeta: ServiceProfileMeta,
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      name + "." + namespace + ".svc." + clusterDomain,
 			Namespace: namespace,
